Improve boot process: move Doom and wallpaper initialization to initial filesystem phase
Moved the logic for downloading and saving Doom game data and default wallpapers from `os-init.js` to `zenfs-init.js`. This ensures all filesystem-related initialization is completed during the early boot phase before the user is prompted to continue, consistent with Wolfenstein 3D setup.

- Integrated Doom and wallpaper loading into `initFileSystem` in `src/system/zenfs-init.js`.
- Added granular progress reporting via the `onProgress` callback for these assets.
- Cleaned up redundant logic and unused imports in `src/system/os-init.js`.
- Verified changes with Playwright and existing E2E tests.
